% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution3d.R
\name{conv3d}
\alias{conv3d}
\title{Convolution 3D for arrays}
\usage{
conv3d(img, kernel)
}
\arguments{
\item{img}{the object to convolve}

\item{kernel}{the kernel}
}
\description{
Convolution 3D for arrays
}
